add post-processing code

+++ update.js
@@ -1,5 +1,7 @@
 const fs = require('fs')
 const cp = require('child_process')
+const github = require('gh-helpers')()
+const { join } = require('path')
 function exec (a, b) {
   console.log('$', [a, b])
   return cp.execSync(a, { ...b, stdio: 'inherit' })
@@ -7,6 +9,10 @@ function exec (a, b) {
 
 function decomp (version) {
   exec('git pull')
+  if (fs.existsSync(join(__dirname, '/.git/refs/heads/client' + version))) {
+    console.log('Already have decompiled', version, 'no work to do')
+    return
+  }
   if (!fs.existsSync('DecompilerMC') || !fs.existsSync('DecompilerMC/.git')) {
     fs.rmSync('DecompilerMC', { recursive: true, force: true })
     exec('git clone http://github.com/extremeheat/DecompilerMC --depth 1')
@@ -25,6 +31,48 @@ function decomp (version) {
   exec(`git push origin client${version}`)
 }
 
+async function postprocess (version, oldVersion, isMajor, oldMajor) {
+  if (isMajor) {
+    console.log('Creating PR as major update from', oldMajor, version)
+    const oldBranch = 'client' + oldMajor
+    const newBranch = 'client' + version
+    const pr = await github.createPullRequest(`MC ${oldMajor} -> ${version}`, `Minecraft Java Edition ${version}<br/>`, oldBranch, newBranch)
+    console.log('Created PR', pr)
+    github.sendWorkflowDispatch({
+      owner: 'extremeheat',
+      repo: 'llm-services',
+      workflow: 'dispatch.yml',
+      branch: 'main',
+      inputs: {
+        action: 'mcpc/reviewUpdateChanges',
+        payload: JSON.stringify({
+          repo: await github.getRepoDetails(),
+          mode: 'release',
+          pr
+        })
+      }
+    })
+  } else {
+    // A snapshot. Get the latest commit SHA on clientlatest and send it over to workflow dispatch so it will diff and make a comment for the commit to clientlatest
+    const sha = fs.readFileSync(join(__dirname, '/.git/refs/heads/clientlatest', 'ascii'))
+    console.log('Snapshot update!', version, 'from', oldVersion, 'Commit SHA:', sha)
+    github.sendWorkflowDispatch({
+      owner: 'extremeheat',
+      repo: 'llm-services',
+      workflow: 'dispatch.yml',
+      branch: 'main',
+      inputs: {
+        action: 'mcpc/reviewUpdateChanges',
+        payload: JSON.stringify({
+          repo: await github.getRepoDetails(),
+          mode: 'snapshot',
+          commit: sha
+        })
+      }
+    })
+  }
+}
+
 async function main () {
   const currentManifest = require('./version_manifest.json')
   const latestManifest = await fetch('https://launchermeta.mojang.com/mc/game/version_manifest.json').then(r => r.json())
@@ -33,12 +81,13 @@ async function main () {
 
   if (currentManifest.latest.release === latestRelease.id && currentManifest.latest.snapshot === latestSnapshot.id) {
     console.log('No new version available')
-    return
   } else {
-    console.log('New version available:', latestRelease.id, latestSnapshot.id)
-    currentManifest.latest.release = latestRelease.id
-    currentManifest.latest.snapshot = latestSnapshot.id
-    fs.writeFileSync('version_manifest.json', JSON.stringify(currentManifest, null, 2))
+    // Store the old versions and check if this is a release update (if latest snap==latest rel). If release update we open PR otherwise we don't for snapshots
+    const oldRel = currentManifest.latest.release
+    const oldSnap = currentManifest.latest.snapshot
+    const isRelease = currentManifest.latest.release === currentManifest.latest.snapshot
+    console.log('New version available:', latestRelease.id, latestSnapshot.id, 'old:', oldSnap)
+    fs.writeFileSync('version_manifest.json', JSON.stringify(latestManifest, null, 2))
     exec('git checkout updator')
     exec('git config user.name "github-actions[bot]"')
     exec('git config user.email "41898282+github-actions[bot]@users.noreply.github.com"')
@@ -46,6 +95,7 @@ async function main () {
     exec(`git commit -m "Update version manifest for ${latestSnapshot.id}"`)
     exec('git push origin updator')
     decomp(latestSnapshot.id)
+    await postprocess(latestSnapshot.id, oldSnap, isRelease, oldRel)
   }
 }
